fix: validate route inputs and harden API response handling
Agent-Logs-Url: https://github.com/piyushahir05/Railway-Route-Optimization-System/sessions/3c850063-0555-4926-b61f-8f95387b8aea

+++ backend/graph_engine/dijkstra.py
@@ -5,7 +5,7 @@ import heapq
 
 def shortest_path(
     graph: dict[str, dict[str, int]], source: str, destination: str
-) -> tuple[int, list[str]]:
+) -> tuple[float, list[str]]:
     """Compute shortest path and total distance using Dijkstra's algorithm."""
     if source not in graph or destination not in graph:
         return float("inf"), []
